msie/requirements.txt
# ---------- Core API ----------
fastapi
uvicorn

# ---------- Data & Analytics ----------
pandas
numpy

# ---------- Environment Variables ----------
python-dotenv

# ---------- Gemini / Google GenAI ----------
google-generativeai
google-genai

# ---------- Validation / Models ----------
pydantic>=2.0

# ---------- Utilities ----------
tenacity
requests
